commit test services

+++ src/main/java/mutant/services/com/domain/model/MutantsCount.java
@@ -3,7 +3,7 @@ package mutant.services.com.domain.model;
 public class MutantsCount {
     private Integer countMutantDna;
     private Integer countHumanDna;
-    private Integer ratio;
+    private double ratio;
 
     public MutantsCount(){}
 }
